+++ robolectric-shadows/shadows-maps/src/test/java/org/robolectric/shadows/maps/ShadowGeocoderTest.java
@@ -1,4 +1,4 @@
-package org.robolectric.shadows;
+package org.robolectric.shadows.maps;
 
 import android.app.Activity;
 import android.location.Address;
